refactor: separate connection string spec tests
